Initial WIP commit for multi region refactor
Essential goal is to spec out the CWE (so that it's not referenced twice
by rules) and then include a new target sns topic that forwards the
event to an external SQS queue.

For the previous cwe_lambda module, we've simply removed the cwe rule
part and included everything else. the CWE details will have to be
passed in as variables.